refactor: seller register 훅/타입 위치 정리
