+++ src/types/seller-register.ts
